Move spec builders to a separate package.
Splits the fork specific builders out into their own classes to avoid SpecConfigBuilder growing bigger all the time.
