+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/config/SpecConfig.java
@@ -19,6 +19,7 @@ import org.apache.tuweni.bytes.Bytes;
 import tech.pegasys.teku.ethereum.execution.types.Eth1Address;
 import tech.pegasys.teku.infrastructure.bytes.Bytes4;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
+import tech.pegasys.teku.spec.config.builder.SpecConfigBuilder;
 
 public interface SpecConfig {
   // Non-configurable constants
